Extract core book table headers & cells into components
