+++ pkg/ui/src/InventoryTable/InventoryTable.svelte
@@ -1,12 +1,11 @@
 <script lang="ts">
-	import { quadIn } from 'svelte/easing';
-
 	import { Checkbox, Button, ButtonColor } from '../';
-
-	import { fadeBgColor } from '../lib/transitions';
 
 	import type { InventoryTableData } from './types';
 	import type { createTable } from './table';
+
+	import CoreTableRow from './CoreTableRow.svelte';
+	import CoreTableHeaders from './CoreTableHeaders.svelte';
 
 	export let table: ReturnType<typeof createTable<InventoryTableData>>;
 
@@ -22,10 +21,8 @@
 		outOfPrint: 'Out of Print'
 	};
 
-	const { tableRow, removeRows, table: tableAction } = table;
+	const { removeRows, table: tableAction } = table;
 	$: ({ rows, selected } = $table);
-
-	const isChecked = (event: Event) => (event?.target as HTMLInputElement)?.checked;
 
 	// table rows + one header row
 	$: rowCount = rows.length + 1;
@@ -41,148 +38,33 @@
 			</div>
 		{/if}
 		<thead>
-			<tr
-				class="whitespace-nowrap"
-				use:tableRow={{
-					position: 0,
-					on: 'change',
-					handleSelect: (event, selected) => {
-						const isSelected = isChecked(event);
-
-						if (isSelected) {
-							selected.set(rows);
-						} else {
-							selected.set([]);
-						}
-					}
-				}}
-			>
-				<th scope="col" class="px-2 text-center">
-					<span class="inline-block">
-						<Checkbox name="Select all" checked={selected.length ? true : false} />
-					</span>
-				</th>
-				<th
-					scope="col"
-					class="table-cell whitespace-nowrap py-4 px-3 text-left text-sm font-medium uppercase tracking-wide text-gray-500"
-				>
-					<span class="hidden lg:inline">{headers.isbn}</span>
-					<span class="inline lg:hidden">book</span>
-				</th>
-				<th
-					scope="col"
-					class="hidden whitespace-nowrap py-4 px-3 text-left text-sm font-medium uppercase tracking-wide text-gray-500 lg:table-cell"
-				>
-					{headers.title}
-				</th>
-				<th
-					scope="col"
-					class="hidden whitespace-nowrap py-4 px-3 text-left text-sm font-medium uppercase tracking-wide text-gray-500 lg:table-cell"
-				>
-					{headers.authors}
-				</th>
-				<th
-					scope="col"
-					class="whitespace-nowrap py-4 px-3 text-left text-sm font-medium uppercase tracking-wide text-gray-500"
-				>
-					{headers.quantity}
-				</th>
-				<th
-					scope="col"
-					class="whitespace-nowrap py-4 px-3 text-left text-sm font-medium uppercase tracking-wide text-gray-500"
-				>
-					{headers.price}
-				</th>
-				<th
-					scope="col"
-					class="hidden whitespace-nowrap py-4 px-3 text-left text-sm font-medium uppercase tracking-wide text-gray-500 sm:table-cell"
-				>
-					{headers.publisher}
-				</th>
+			<CoreTableHeaders {table} {headers} let:headers={{ publisher, editedBy, outOfPrint }}>
 				<th
 					scope="col"
 					class="hidden whitespace-nowrap py-4 px-3 text-left text-sm font-medium uppercase tracking-wide text-gray-500 md:table-cell"
 				>
-					{headers.year}
+					{publisher}
 				</th>
 				<th
 					scope="col"
 					class="hidden whitespace-nowrap py-4 px-3 text-left text-sm font-medium uppercase tracking-wide text-gray-500 xl:table-cell"
 				>
-					{headers.editedBy}
+					{editedBy}
 				</th>
 				<th
 					scope="col"
 					class="hidden whitespace-nowrap py-4 px-3 text-left text-sm font-medium uppercase tracking-wide text-gray-500 xl:table-cell"
 				>
-					{headers.outOfPrint}
+					{outOfPrint}
 				</th>
-			</tr>
+			</CoreTableHeaders>
 		</thead>
 
 		<tbody>
 			{#each rows as row (row.key)}
-				{@const { isbn, title, authors, quantity, price, publisher, year, editedBy, outOfPrint, rowIx } = row}
-				<tr
-					in:fadeBgColor={{ duration: 200, easing: quadIn, color: 'rgb(220 252 231)' }}
-					out:fadeBgColor={{ duration: 150, easing: quadIn, color: 'rgb(254 226 226)' }}
-					use:tableRow={{
-						// Header row starts the count at 0
-						position: rowIx + 1,
-						on: 'change',
-						handleSelect: (event, selected) => {
-							const isSelected = isChecked(event);
-
-							if (isSelected) {
-								selected.update((rows) => [...rows, row]);
-							} else {
-								selected.update((rows) => rows.filter((r) => r.key !== row.key));
-							}
-						}
-					}}
-					class={`whitespace-nowrap text-sm font-light text-gray-500 ${
-						selected.includes(row) ? 'bg-gray-100' : 'even:bg-gray-50'
-					}`}
-				>
-					<td
-						class={`px-2 text-center sm:align-middle border-l-4 
-							${selected.includes(row) ? 'border-teal-500' : 'border-transparent'}
-						`}
-					>
-						<span class="inline-block">
-							<Checkbox name={`Select ${title}`} checked={selected.includes(row)} />
-						</span>
-					</td>
-					<th scope="row" class="py-4 px-3 text-left font-medium text-gray-800 lg:w-auto lg:max-w-none">
-						{isbn}
-						<dl class="font-normal lg:hidden">
-							<dt class="sr-only">Title:</dt>
-							<dd class="mt-1 truncate font-light text-gray-500">{title}</dd>
-							<dt class="sr-only lg:hidden">Authors:</dt>
-							<dd class="mt-1 truncate font-light text-gray-500 lg:hidden">{authors}</dd>
-							<dt class="sr-only md:hidden">Year:</dt>
-							<dd class="mt-1 truncate font-light text-gray-500 md:hidden">{year}</dd>
-						</dl>
-					</th>
-					<td class="hidden px-3 py-4 lg:table-cell">
-						{title}
-					</td>
-					<td class="hidden py-4 px-3 lg:table-cell">
-						{authors}
-					</td>
-					<td class="py-4 px-3 text-left">
-						<span class="rounded-md bg-gray-100 px-3 py-2">
-							{quantity}
-						</span>
-					</td>
-					<td class="py-4 px-3 text-left">
-						{price}
-					</td>
-					<td class="hidden py-4 px-3 sm:table-cell">
+				<CoreTableRow {table} {row} let:row={{ title, publisher, editedBy, outOfPrint }}>
+					<td class="hidden py-4 px-3 md:table-cell">
 						{publisher}
-					</td>
-					<td class="hidden py-4 px-3 text-left md:table-cell">
-						{year}
 					</td>
 					<td class="hidden py-4 px-3 xl:table-cell">
 						{editedBy}
@@ -192,7 +74,7 @@
 							<Checkbox name={`${title} is out of print: ${outOfPrint}`} checked={outOfPrint} disabled />
 						</span>
 					</td>
-				</tr>
+				</CoreTableRow>
 			{/each}
 		</tbody>
 	</table>
